# hugoproxy/proxy/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
    "/geocode": {
      "post": {
        "description": "This endpoint processes geocoding requests and returns the result.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "GeoCoding"
        ],
        "summary": "Handle geocoding",
        "parameters": [
          {
            "description": "Geocoding Query",
            "name": "query",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/controller.geoCodeUnm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully processed geocoding",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "description": "This endpoint authenticates a user and returns a JWT token.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Authentication"
        ],
        "summary": "User Login",
        "parameters": [
          {
            "description": "User Credentials",
            "name": "credentials",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/service.Credentials"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JWT Token",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "string"
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/search": {
      "post": {
        "description": "This endpoint processes search requests and returns search results.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Search"
        ],
        "summary": "Search",
        "parameters": [
          {
            "description": "Search Query",
            "name": "query",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/controller.searchUnm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully processed search request",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "controller.geoCodeUnm": {
      "type": "object",
      "properties": {
        "query": {
          "type": "string"
        }
      }
    },
    "controller.searchUnm": {
      "type": "object",
      "properties": {
        "query": {
          "type": "string"
        }
      }
    },
    "service.Credentials": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    }
  }
}